+++ app/project/bootstrappers/console/commands/Commands.php
@@ -6,7 +6,7 @@
  */
 namespace Project\Bootstrappers\Console\Commands;
 use RDev\Applications\Bootstrappers\Bootstrapper;
-use RDev\Console\Commands\Commands as ConsoleCommands;
+use RDev\Console\Commands\CommandCollection;
 use RDev\IoC\IContainer;
 
 class Commands extends Bootstrapper
@@ -14,17 +14,17 @@ class Commands extends Bootstrapper
     /**
      * Sets the console commands from this project
      *
-     * @param ConsoleCommands $commands The commands to add to
+     * @param CommandCollection $commandCollection The commands to add to
      * @param IContainer $container The dependency injection container
      */
-    public function run(ConsoleCommands $commands, IContainer $container)
+    public function run(CommandCollection $commandCollection, IContainer $container)
     {
         $commandClasses = require $this->paths["configs"] . "/console/commands.php";
 
         // Instantiate each command class
         foreach($commandClasses as $commandClass)
         {
-            $commands->add($container->makeShared($commandClass));
+            $commandCollection->add($container->makeShared($commandClass));
         }
     }
 }